+++ lib/template.rb
@@ -1,5 +1,5 @@
 class Template
-  VERSION = Gem::Version.new("0.2.1")
+  VERSION = Gem::Version.new("0.2.2")
 
   def initialize(input, io: StringIO.new)
     @input = input