+++ lib/template/parser/template.rb
@@ -15,10 +15,14 @@ class Template
       end
 
       rule(:code_part) do
-        left_curly_bracket.ignore >> code >> (right_curly_bracket.ignore | any.absent?)
+        left_curly_bracket.ignore >> code >>
+          (right_curly_bracket.ignore | any.absent?)
       end
 
-      rule(:template) { (text_part.as(:text) | code_part.as(:code)).repeat(1) }
+      rule(:template) do
+        (text_part.as(:text) | code_part.as(:code)).repeat(1) |
+          str("").as(:text).repeat(1, 1)
+      end
 
       root(:template)
     end