+++ spec/code/parser/call_spec.rb
@@ -21,9 +21,7 @@ RSpec.describe Code::Parser::Call do
               },
             },
           },
-          right: [{
-            name: "times",
-          }],
+          right: [{ name: "times" }],
         },
       },
     ],