+++ CHANGELOG.md
@@ -5,9 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.2.2 / 2022-08-31
+
+- Fix parsing error when the template is empty, e.g. ""
+
 ## 0.2.1 / 2022-08-31
 
-- Fix error on empty code like `Hello {`
+- Fix parsing error on empty code like `Hello {`
 
 ## 0.2.0 / 2022-08-30
 
@@ -16,4 +20,3 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## 0.1.0 / 2022-07-28
 
 - Initial version with interpolation
-